@@ -12,15 +12,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.12",
+    "@angular/animations": "^18.2.13",
     "@angular/cdk": "^17.3.10",
-    "@angular/common": "^17.3.12",
-    "@angular/compiler": "^17.3.12",
-    "@angular/core": "^17.3.12",
-    "@angular/forms": "^17.3.12",
-    "@angular/platform-browser": "^17.3.12",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/router": "^17.3.12",
+    "@angular/common": "^18.2.13",
+    "@angular/compiler": "^18.2.13",
+    "@angular/core": "^18.2.13",
+    "@angular/forms": "^18.2.13",
+    "@angular/platform-browser": "^18.2.13",
+    "@angular/platform-browser-dynamic": "^18.2.13",
+    "@angular/router": "^18.2.13",
     "@ng-bootstrap/ng-bootstrap": "^16.0.0",
     "@ng-idle/core": "^13.0.0",
     "@popperjs/core": "^2.11.8",
@@ -46,14 +46,14 @@
     "zone.js": "~0.14.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.11",
+    "@angular-devkit/build-angular": "^18.2.12",
     "@angular-eslint/builder": "^17.5.2",
     "@angular-eslint/eslint-plugin": "^17.5.2",
     "@angular-eslint/eslint-plugin-template": "^17.5.2",
     "@angular-eslint/schematics": "^17.5.2",
     "@angular-eslint/template-parser": "^17.3.12",
-    "@angular/cli": "^17.3.11",
-    "@angular/compiler-cli": "^17.3.12",
+    "@angular/cli": "^18.2.12",
+    "@angular/compiler-cli": "^18.2.13",
     "@types/jasmine": "~4.6.4",
     "@types/lodash": "^4.17.13",
     "@types/markdown-it": "^13.0.7",
