@@ -1,9 +1,10 @@
 import { NO_ERRORS_SCHEMA } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 
 import { InformationModalComponent } from './information-modal.component';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('InformationModalComponent', () => {
   let component: InformationModalComponent;
@@ -11,14 +12,16 @@ describe('InformationModalComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [HttpClientTestingModule, InformationModalComponent],
+      schemas: [NO_ERRORS_SCHEMA],
+      imports: [InformationModalComponent],
       providers: [
         {
           provide: NgbActiveModal,
           useValue: NgbActiveModal,
         },
+        provideHttpClient(withInterceptorsFromDi()),
+        provideHttpClientTesting(),
       ],
-      schemas: [NO_ERRORS_SCHEMA], // only for directives
     }).compileComponents();
 
     fixture = TestBed.createComponent(InformationModalComponent);
