@@ -1,8 +1,9 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 
 import { PagesTopnavbarComponent } from './pages-topnavbar.component';
 import { NO_ERRORS_SCHEMA } from '@angular/core';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('PagesTopnavbarComponent', () => {
   let component: PagesTopnavbarComponent;
@@ -10,8 +11,9 @@ describe('PagesTopnavbarComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [HttpClientTestingModule, PagesTopnavbarComponent],
       schemas: [NO_ERRORS_SCHEMA],
+      imports: [PagesTopnavbarComponent],
+      providers: [provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()],
     }).compileComponents();
 
     fixture = TestBed.createComponent(PagesTopnavbarComponent);
