@@ -1,7 +1,8 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 
 import { WebsocketTestComponent } from './websocket-test.component';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('WebsocketTestComponent', () => {
   let component: WebsocketTestComponent;
@@ -9,7 +10,8 @@ describe('WebsocketTestComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [WebsocketTestComponent, HttpClientTestingModule],
+      imports: [WebsocketTestComponent],
+      providers: [provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()],
     }).compileComponents();
 
     fixture = TestBed.createComponent(WebsocketTestComponent);
